Replace firewood-connection with client/server implementation
This diff adds support for an async DB trait, shown in api.rs.

The general idea is to create a channel to the thread running the DB,
rather than passing the DB object around. For batches, each batch gets
an ID, but the batch itself stays server side.

The following API calls are fully implemented:
 * set_code
 * set_nonce
 * set_state
 * create_account
 * no_root_hash
 * root_hash

Also fixed up the shutdown code, by implementing Drop on Connection.